+++ Changes.md
@@ -1,5 +1,9 @@
+Release 11.1
+============
+
 * #450 - Refactor allteams.jsp
 * Fix bug where the datasource doesn't get initialized before the setup page is loaded
+
 
 Release 11.0
 ============